requirements.txt
streamlit
google-generativeai
pypdf
python-pptx
youtube-transcript-api
gTTS
graphviz